[Login] Updated to 3.0.3.0 r330995;

+++ AAEmu.Login/Core/Packets/C2L/CARequestReconnectPacket.cs
@@ -13,8 +13,8 @@ public class CARequestReconnectPacket : LoginPacket
     {
         var pFrom = stream.ReadUInt32();
         var pTo = stream.ReadUInt32();
-        var accountId = stream.ReadUInt32();
-        var gsId = stream.ReadByte();
+        var accountId = stream.ReadUInt64();
+        var gsId = stream.ReadByte(); // wid
         var cookie = stream.ReadUInt32();
         var macLength = stream.ReadUInt16();
         var mac = stream.ReadBytes(macLength);
